Added rmjunk script.
Find all messages in the Maildir ~/Mail/Junk, uses bogofilter to
classify them as spam, then deletes them. If you pass a parameter, it is
used as the argument to find -mtime to select messages older than some
number of days. For example, use rmjunk +7 to delete massages older than
7 days.